+++ mono/coder/serializers.py
@@ -1,7 +1,23 @@
 """Coder's serializers"""
 from rest_framework import serializers
 
-from .models import Snippet
+from .models import Snippet, Tag
+
+
+class TagSerializer(serializers.ModelSerializer):
+    """Tag serializer"""
+    created_by = serializers.HiddenField(
+        default=serializers.CurrentUserDefault()
+    )
+
+    class Meta:
+        model = Tag
+        fields = [
+            'id',
+            'name',
+            'color',
+            'created_by',
+        ]
 
 
 class SnippetSerializer(serializers.ModelSerializer):
@@ -9,6 +25,7 @@ class SnippetSerializer(serializers.ModelSerializer):
     created_by = serializers.HiddenField(
         default=serializers.CurrentUserDefault()
     )
+    tags = TagSerializer(many=True, read_only=True)
 
     class Meta:
         model = Snippet
@@ -21,6 +38,7 @@ class SnippetSerializer(serializers.ModelSerializer):
             'public',
             'public_id',
             'created_by',
+            'tags',
         ]
         read_only_fields = [
             'html',